refactor(utils): unify captureSnapshot + trace into lib/utils/trace.js
Both files belonged to the same flow (capture artifacts -> render trace.md).
Splitting them across two files made the import sites awkward, so they're
now consolidated:

- `lib/utils/captureSnapshot.js` deleted; its `captureSnapshot`,
  `normalizeBrowserLogs`, and `captureStorageState` moved into
  `lib/utils/trace.js`.
- `writeTraceMarkdown` (was inlined in `bin/mcp-server.js`) moved into
  `lib/utils/trace.js` too — it's the same trace-record concern.
- All three call sites (`bin/mcp-server.js`, `lib/plugin/aiTrace.js`,
  `lib/plugin/pageInfo.js`) now import from a single module.
- Stale `writeFileSync` import dropped from `bin/mcp-server.js`.

72 unit tests still passing.
