+++ lib/plugin/aiTrace.js
@@ -8,8 +8,7 @@ import recorder from '../recorder.js'
 import event from '../event.js'
 import output from '../output.js'
 import { deleteDir, clearString } from '../utils.js'
-import { captureSnapshot } from '../utils/captureSnapshot.js'
-import { pickActingHelper, traceDirFor, artifactLinks } from '../utils/trace.js'
+import { captureSnapshot, pickActingHelper, traceDirFor, artifactLinks } from '../utils/trace.js'
 
 const defaultConfig = {
   deleteSuccessful: false,
