+++ lib/plugin/pageInfo.js
@@ -4,8 +4,7 @@ import Container from '../container.js'
 import recorder from '../recorder.js'
 import event from '../event.js'
 import { scanForErrorMessages } from '../html.js'
-import { captureSnapshot } from '../utils/captureSnapshot.js'
-import { pickActingHelper } from '../utils/trace.js'
+import { captureSnapshot, pickActingHelper } from '../utils/trace.js'
 import { output } from '../index.js'
 import store from '../store.js'
 import { humanizeString, ucfirst } from '../utils.js'
